requirements.txt
celery==4.4.7
django>=4.0
django-cors-headers==3.6.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.21.0
flower==0.9.7
psycopg2-binary
redis==3.5.3